requirements.txt
streamlit
PyMuPDF
sentence-transformers
faiss-cpu
numpy
google-generativeai
tqdm
ebooklib
bs4
lxml
